% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date9.R
\name{as_date9}
\alias{as_date9}
\title{Convert a vector to date9 character string.}
\usage{
as_date9(x)
}
\arguments{
\item{x}{Date, date-time or character vector.}
}
\value{
A character vector in date9 format.
}
\description{
date9 is a SAS format of the form 03APR1997 - two-character
  date, then a capitalized 3-character month string, then 4-digit year.  This
  function coerces the input to a date (if a date-time or string are input,
  for example) using lubridate::as_date(), the it formats the date into a
  date9 string.
}
\examples{
as_date9(Sys.Date())
as_date9(Sys.time())
}
