% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmisc.R
\docType{package}
\name{cfmisc}
\alias{cfmisc}
\title{\code{cfmisc} package}
\description{
Miscellaneous functions and utilities for Cystic Fibrosis research.
}
\details{
For a general overview of what this package covers please see
  \code{browseVignettes("cfmisc")}
}
