% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stat.R
\name{est_int_str}
\alias{est_int_str}
\title{Estimate and (confidence) interval string}
\usage{
est_int_str(
  est,
  lower,
  upper,
  est_digits = 2,
  ci_digits = est_digits,
  na = "",
  plus_prefix = T
)
}
\arguments{
\item{est}{Estimate, a numeric vector}

\item{lower}{Lower bound of the interval}

\item{upper}{Upper bound of the interval}

\item{est_digits}{Number of digits to round to for the estimate (default =
2).}

\item{ci_digits}{Number of digits to round to for the CI (lower and upper).
Defaults to est_digits.}

\item{na}{String to return if estimate is NA (optional - NULL for no action).}

\item{plus_prefix}{logical.  Add a plus sign to positive estimates?}
}
\description{
Creates a string representation of 3 numbers:  Point estimate,
  lower bound for an interval, upper bound for an interval.  The purpose is
  displaying statistical results (prediciton intervals, confidence intervals,
  etc).  The default number of digits comes from the NEJM guidelines, which
  recommend two digits as a good starting point for association measures
  \url{https://www.nejm.org/author-center/new-manuscripts}
}
\examples{
est_int_str(c(1,2), c(3,4), c(5.6, 7.8))
est_int_str(c(NA,2), c(3,4), c(5.6, 7.8), est_digits = 0, plus_prefix = TRUE, na = "miss")
}
