% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{form_f}
\alias{form_f}
\title{Format f}
\usage{
form_f(x, digits = 0)
}
\arguments{
\item{x}{A numeric value.}

\item{digits}{The desired number of digits after the decimal point.}
}
\description{
Format C with format = f - simple wrapper.
}
\examples{
form_f(293.293893, 2)
}
