% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_label.R
\name{glimpse_label}
\alias{glimpse_label}
\title{Glimpse label}
\usage{
glimpse_label(x, width = NULL, ...)
}
\arguments{
\item{x}{A dataframe, hopefully with some label attributes.}

\item{width}{The character width of the printout.  Default: automatic guess.}

\item{...}{Not used.  Copy from glimpse().}
}
\description{
Similar to glimpse from the tibble package, with labels printed
  instead of values.
}
\details{
Labels are a customary part of SAS datasets.  When we use
  \code{haven::read_sas()}, these labels are read in and stored as attributes
  using \code{labeller}. These are useful in understanding datasets, and with
  other packages like \code{gtsummary}, so this function was created to look
  at them quickly.
}
