% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcf_df.R
\name{lpcf_df}
\alias{lpcf_df}
\title{Last percentile carried forward (for dataframes)}
\usage{
lpcf_df(
  dat,
  age,
  height,
  sex,
  group_vars = NULL,
  output_name = "projected_height"
)
}
\arguments{
\item{dat}{A dataframe or tibble.  Grouping will be ignored.}

\item{age}{Unquoted variable name in \code{dat} containing ages (years).}

\item{height}{Unquoted variable name in \code{dat} containing height (cm).}

\item{sex}{Unquoted variable name in \code{dat} containing sex at birth ("M" or "F").  Sex at birth should not change for participants.}

\item{group_vars}{A character vector containing the grouping variable(s) for projection, usually the single column containing subject id's.  In hierarchical studies this could be a vector or grouping variables like c("country", "region", "subject").}

\item{output_name}{A character vector for the name of the new column containing the projected heights.}
}
\value{
A tibble with a column added for projected height.  Grouping will be dropped and the dataframe will be sorted by grouping vars, then age.
}
\description{
This function takes a set of clinical observations and projects
   the heights of subjects forward so that their z score stays the same in
   the case of missing values.
}
\examples{
test_dat <- data.frame(subject = c("A1","A1","B2","B2"), sex = rep("F", 4),
    visit = c(1,2,1,2), visit_age = c(12, 12.4, 14, 14.2),
    ht = c(140, NA, 135, NA))
lpcf_df(test_dat, age = visit_age, height = ht, sex = sex,
        group_vars = "subject")
}
