% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stat.R
\name{n_pct_str}
\alias{n_pct_str}
\title{n percent string}
\usage{
n_pct_str(n, d, show_d = F, digits = 0, na = "")
}
\arguments{
\item{n}{Numerator vector (whole numbers only)}

\item{d}{Denominator vector (whole numbers only)}

\item{show_d}{T/F - show the denominator in the string?}

\item{digits}{Digits after the decimal point to show.}

\item{na}{String to return if n is missing.  If null then the output not
modified.}
}
\description{
Creates a string (glue) output, given a numerator vector (n) and
  a denominator vector (d).  The percentage and n are always shown, and the
  denominator is shown only if show_d = T.
}
\examples{
n_pct_str(c(5,6), c(7,7))
n_pct_str(c(NA,6), c(7,7), show_d = TRUE, digits = 2, na = "(Missing)")
}
