% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ht.R
\name{project_ht}
\alias{project_ht}
\title{Project heights}
\usage{
project_ht(sex, old_height, old_age, new_age)
}
\arguments{
\item{sex}{Sex at birth, "M" for male and "F" for female.}

\item{old_height}{Height to be projected.}

\item{old_age}{Age at \code{old_height}.}

\item{new_age}{Age we're projecting to (where we guess height).}
}
\value{
Returns a numeric vector with height guesses at \code{new_age}.
}
\description{
Project a vector of heights using the CDC growth charts.
}
\details{
The motivation for this function is to project previously measured
heights forward in order to have a reasonable guess at a new timepoint.
This explicitly assumes that a person would have grown at a rate which
would place them in the same percentile (or z score), as compared to the
CDC growth charts.

The advantages of this approach are
\enumerate{
\item it requires no data aside from one
previous height measurement, and could be used for individual subjects if
needed, and
\item it is conservative compared to last observation carried
forward (LOCF) if the height is to be used in GLI spirometry equations
later (to calculate ppFEV1, for example).
}

On the other hand, LOCF is more conservative in analysis of height
as an outcome, so we strongly recommend its use over this method when
height is the target outcome.  Data-driven imputation methods may have
better performance in cohorts which differ significantly from the CDC's
cohort, or for projecting over large time periods.
}
\examples{
# for a male subject whose height was 152cm at 14 years old and
# is missing the height measurement at age 14.6:
project_ht(sex = "M", old_height = 152, old_age = 14, new_age = 14.6)
}
