% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stat.R
\name{pval_nejm}
\alias{pval_nejm}
\title{NEJM-style P value formatter}
\usage{
pval_nejm(p, na = "")
}
\arguments{
\item{p}{numeric vector, all in [0,1] or NA.}

\item{na}{String to return if missing.  Empty string is default.}
}
\value{
A character string representing a P value.
}
\description{
Updated NEJM guildines state: "In general, P values larger than
  0.01 should be reported to two decimal places, and those between 0.01 and
  0.001 to three decimal places; P values smaller than 0.001 should be
  reported as P<0.001."
  (\url{https://www.nejm.org/author-center/new-manuscripts})
}
\examples{
pval_nejm(c(0.2, 0.000003, NA))
}
