% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_sas.R
\name{rc_crf}
\alias{rc_crf}
\alias{clean_crf}
\alias{read_crf}
\title{Read and cleanup iMedidata CRF data}
\usage{
clean_crf(sas_df, vn = F, date = T)

read_crf(file, vn = F, date = T)
}
\arguments{
\item{sas_df}{A dataframe.}

\item{vn}{logical - should a visit number be parsed from the \code{folder}
column?}

\item{date}{logical - should all date-time columns be converted to dates?}

\item{file}{File path to a SAS dataset, passed on to haven::read_sas().}
}
\description{
CRF datasets from iMedidata Rave have some common, annoying
  features when loaded in.  These include (1) all caps variable names (2)
  character columns having blanks instead of NA values (3) date-time objects
  instead of dates.  This function is a convenience wrapper that fixes these,
  and optionally adds a visit number column based on the Folder.
}
\details{
\code{clean_crf()} works on already-loaded datasets.
  \code{read_crf()} loads the dataset using haven given a filepath, then
  applies \code{clean_crf()}.
}
\examples{
mtcars \%>\% clean_crf
}
