% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity_fev.R
\name{schluchter_severity}
\alias{schluchter_severity}
\alias{severity_schluchter}
\alias{severity_cutoff_schluchter}
\title{Disease severity (Schluchter, 2005)}
\usage{
severity_schluchter(age, ppfev, type = "integer_age", factorize = T)

severity_cutoff_schluchter(age, line = "severe", type = "integer_age")
}
\arguments{
\item{age}{Subject age in years.}

\item{ppfev}{Percentage of predicted FEV1 (Schluchster used Knudson eq.)}

\item{type}{Either "integer_age" or "linear_interpolation" for smooth function.}

\item{factorize}{T or F - return a factor?}

\item{line}{Either "severe" or "mild".}
}
\description{
Disease severity based on ppFEV1 given age.  Used in cystic fibrosis.
}
\details{
Severity is a measure of disease aggressiveness based on age and ppFEV.  Essentially the observed ppFEV is compared with quantile-based cutoffs for that age.  This makes severity distinct from stage or condition, which usually just uses ppFEV1 alone (e.g. <40 ppFEV1 is on definition of an advanced stage).

Schluchter identifies 'mild' and 'severe' phenotypes.  We label the group between these moderate.
Data comes from Schluchter (2006) "Classifying severity...", DOI: 10.1164/rccm.200512-1919OC    Table E1.
}
