% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sev_tab_schluchter}
\alias{sev_tab_schluchter}
\title{Schluchter's FEV severity equations.}
\format{
A data frame with 27 rows and 3 variables: \describe{
  \item{age_lb}{Age lower bound}
  \item{severe}{The lower bound to be classified as severe.}
  \item{mild}{The upper bound to be classified as mild.}
   ... }
}
\source{
\url{doi.org/10.1164/rccm.200512-1919OC}
}
\usage{
sev_tab_schluchter
}
\description{
Data comes from Schluchter (2006) "Classifying severity...", DOI:
  10.1164/rccm.200512-1919OC
}
\keyword{datasets}
