% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uspi_2020dec20_invitro}
\alias{uspi_2020dec20_invitro}
\title{A list of alleles from the United States Prescribing Info which are
  responsive to ETI in vitro.}
\format{
A character vector with 178 variant names (some are compound alleles).
}
\source{
See table 4 in
  \url{https://www.accessdata.fda.gov/drugsatfda_docs/label/2020/212273s002lbl.pdf}.
  If this URL breaks then search "trikafta" in the 'access data FDA' page.
}
\usage{
uspi_2020dec20_invitro
}
\description{
Prescribing info in the United States was updated to include
  alleles which were shown to be responsive to ETI in vitro on December 20,
  2020.  The primary importance of this change is FDA-approved eligibility
  for subjects with these mutations but zero copies of F508del.
}
\keyword{datasets}
