% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_445_102}
\alias{vx_445_102}
\title{VX-445-102 variant list.}
\format{
A data frame with 220 rows and 10 variables: \describe{
  \item{variant}{The 'mutation' name, which is typically the CF legacy name}
  \item{category}{ The full category name as stated on the sheet for example
  "Nonsense mutations"}
  \item{cat_abbr}{An abbreviated category name for easier typing, e.g. "nonsense"}
  }
}
\source{
\url{https://www.cff.org/PDF-Archive/Study-VX-445-102-Eligible-Mutations-April-2018.pdf}
}
\usage{
vx_445_102
}
\description{
A list of variants to identify "minimal function" mutations for
  eligibility in the F508del heterozygous clinical trial of ETI therapy (
  doi.org/10.1056/NEJMoa1908639)
}
\keyword{datasets}
